+++ packages/ui/src/form/primitives/select/item.tsx
@@ -4,6 +4,10 @@ import { forwardRef } from '@wordpress/element';
 import { check } from '@wordpress/icons';
 import itemPopupStyles from '../../../utils/css/item-popup.module.css';
 import resetStyles from '../../../utils/css/resets.module.css';
+import {
+	getItemPopupSizeClassName,
+	normalizeItemPopupSize,
+} from '../../../utils/item-popup-size';
 import { Icon } from '../../../icon';
 import type { SelectItemProps } from './types';
 
@@ -12,12 +16,14 @@ export const Item = forwardRef< HTMLDivElement, SelectItemProps >(
 		{ className, value, size = 'default', children, ...restProps },
 		ref
 	) {
+		const normalizedSize = normalizeItemPopupSize( size );
+
 		return (
 			<_Select.Item
 				className={ clsx(
 					resetStyles[ 'box-sizing' ],
 					itemPopupStyles.item,
-					itemPopupStyles[ `is-size-${ size }` ],
+					getItemPopupSizeClassName( size, itemPopupStyles ),
 					className
 				) }
 				value={ value }
@@ -30,7 +36,7 @@ export const Item = forwardRef< HTMLDivElement, SelectItemProps >(
 						itemPopupStyles[ 'item-icon' ],
 						itemPopupStyles[ 'item-indicator-icon' ]
 					) }
-					size={ size === 'small' ? 20 : 24 }
+					size={ normalizedSize === 'small' ? 20 : 24 }
 				/>
 				<_Select.ItemText>{ children }</_Select.ItemText>
 			</_Select.Item>